Add files via upload
The company wishes to assess the impact of its marketing strategy over the past 2 years for its flagship drug NZT-48, which is used to treat chronic Lyme disease. Your job as a data scientist is to estimate a dollar-value of sales attributable to each strategy. Your findings will be used to guide future budget-setting. 
Additionally, the company wants to know the extent of loss of potential sales due to a new competitor drug entrance into the market since mid-2015. While the new drug is not a perfect substitute for NZT-48, the company is concerned about its impact on sales and is considering the costly option of switching marketing strategies.